extract util functions from MLocationBody
